+++ packages/workbench/src/material/index.ts
@@ -1,2 +1,5 @@
-export * from './basic';
-export * from './pbr';
+export * from './basic-material';
+export * from './pbr-material';
+
+export * from './shader-flat-material';
+export * from './shader-lit-material';